[Feat] PostCell 관련 UI constants 별도로 관리하던 폴더 제거 및 각각의 subview에 constants 모델 추가
